chore: extract MultiByteNote to testutil package
